First release
First release of the speaker schematic.  Believed to be complete and correct, no further changes planned.

First release for the amp schematic.  For the portion of the amp shown (power input), believed to be correct with the exception of the right half of the power switch, which is incomplete and might be incorrect.  No further updates currently planned.